fixed all load issues/settings
any script load issues are now FIXED
settings menu is now added and it shows you the data path/scene icon color

you can't change settings yet

+++ addons/TemplateSaver/Code/TemplateData.gd
@@ -1,6 +1,7 @@
 @tool
 extends Node
 
+var DataPath = ""
 var Data: Dictionary = {}
 
 func Add(Name: String, ID: String, Template: PackedScene, UI: HBoxContainer) -> bool:
@@ -16,8 +17,9 @@ func Add(Name: String, ID: String, Template: PackedScene, UI: HBoxContainer) -> 
 	return false
 
 func Remove(Name: String) -> bool:
-	if Data.has(Name):
-		var Directory: DirAccess = DirAccess.open("C:/Godot/")
+	if Data.has(Name) and get_tree().root.has_node("TemplateData"):
+		var TreeTemplateData: Object = get_tree().root.get_node("TemplateData")
+		var Directory: DirAccess = DirAccess.open(TreeTemplateData.DataPath)
 		
 		Directory.remove(Data[Name].ID + ".dat")
 		Data[Name].UI.queue_free()
